+++ src/main/resources/fr/insalyon/heptabits/pldagile/hello-view.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="fr.insalyon.heptabits.pldagile.HelloController">
+      fx:controller="fr.insalyon.heptabits.pldagile.controller.HelloController">
   <padding>
     <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
   </padding>
